+++ flutter/lib/src/integrations/load_debug_images_integration.dart
@@ -1,0 +1,2 @@
+export 'native_load_debug_images_integration.dart'
+    if (dart.library.js_interop) 'web_load_debug_images_integration.dart';